test: Add Java 17 to the matrix build
When trying to run the test with Java 17, there was an issue with assertion accessing internal API. As reported in: https://github.com/google/compile-testing/issues/222.

As a fix, replaced asserts with plain text asserts, inspired by https://github.com/google/auto/pull/1004

To be running the test on Java 17. Need to update the processor test to