feat: Support properties loading for alternate language codes (iw, in, ji) in Java < 17
iw, in, ji are obsolete codes that are used by Java < 17. While new codes (he, id, yi) are accepted, all API calls on the locale object always return the old codes (regardless of the input tag). This means that properties files must use the old codes in their filename to be loaded as a ResourceBundle.

The locale object behavior was changed in Java 17+: locale object now return the new codes and that regardless of the input tag used. Additionally, On the properties file loading, Java 17+ has some logic to attempt reading both the new and the old locale codes.

This change adds for Java 8 up to 17, the ability to read from both the new and old tag.

More concretely, before the change the file for Hebrew must have named "Messages_iw.properties" to be loaded, but now it can be either "Messages_iw.properties" or "Messages_he.properties".

Details:
- Add simple utility to list alternate language codes, and to build alternate resource name
- CharsetDecoderResourceBundleControl change try loading using the alternate locale code / resource name
- DEFAULT_OR_BACKPORT ResourceBundleProvider is updated to use the L10nMessages specific ResourceBundle.Control. it could be a minor change of behavior for Java 9-11 but probably unnoticeable